CSE (C-code) refactoring
